requirements.txt
docker>=3.7.2
graphviz>=0.10.0
ruamel.yaml>=0.15.94
jsonschema>=3.0
dnspython>=1.15
webdavclient>=1.0.8
